+++ tests/browser/test_dashboard.py
@@ -1,0 +1,125 @@
+import random
+from unittest import mock
+
+import allure
+import pytest
+from selenium.webdriver.common.keys import Keys
+
+from core import helpers
+from directory_constants import choices
+from tests.browser.common_selectors import (
+    DashboardContents,
+    DashboardContentsOnSuccess,
+    DashboardContentsWithoutSuccess,
+    DashboardModalLetsGetToKnowYou,
+    HeaderSignedIn,
+)
+from tests.browser.util import (
+    attach_jpg_screenshot,
+    find_element,
+    should_not_see,
+    should_not_see_errors,
+    should_see_all_elements,
+)
+from tests.helpers import create_response
+
+pytestmark = pytest.mark.browser
+
+
+@allure.step('Enter sectors user is interested in: {industries}')
+def submit_industries(browser, industries):
+    industries_input = find_element(
+        browser, DashboardModalLetsGetToKnowYou.INDUSTRIES_INPUT
+    )
+    for industry in industries:
+        industries_input.send_keys(industry)
+        industries_input.send_keys(Keys.ENTER)
+
+    attach_jpg_screenshot(
+        browser,
+        'After entering industries',
+        selector=DashboardModalLetsGetToKnowYou.MODAL
+    )
+
+    continue_button = find_element(browser, DashboardModalLetsGetToKnowYou.SUBMIT)
+    continue_button.click()
+
+
+@pytest.mark.django_db
+@mock.patch.object(helpers, 'get_dashboard_export_opportunities')
+@mock.patch.object(helpers, 'get_dashboard_events')
+@mock.patch.object(helpers, 'create_company_profile')
+def test_dashboard_with_success_query_parameter(
+    mock_create_company_profile, mock_get_dashboard_events,
+    mock_get_dashboard_export_opportunities, mock_get_company_profile,
+    server_user_browser_dashboard, single_event, single_opportunity
+):
+    def side_effect(_):
+        mock_get_company_profile.return_value = {
+            'expertise_countries': [],
+            'expertise_industries': ['SL10001', 'SL10002'],
+        }
+
+    mock_get_dashboard_events.return_value = create_response()
+    mock_get_dashboard_events.side_effect = [[], [single_event], [single_event]]
+    mock_get_dashboard_export_opportunities.return_value = create_response()
+    mock_get_dashboard_export_opportunities.side_effect = [
+        [], [single_opportunity], [single_opportunity]
+    ]
+    mock_create_company_profile.return_value = create_response()
+    mock_create_company_profile.side_effect = side_effect
+    live_server, user, browser = server_user_browser_dashboard
+    should_not_see_errors(browser)
+
+    should_see_all_elements(browser, DashboardModalLetsGetToKnowYou)
+
+    sector_labels = [label for _, label in choices.SECTORS]
+    industries = random.sample(sector_labels, random.randint(1, 5))
+    submit_industries(browser, industries)
+
+    attach_jpg_screenshot(browser, 'Dashboard')
+    should_not_see(browser, DashboardModalLetsGetToKnowYou)
+    should_see_all_elements(browser, HeaderSignedIn)
+    should_see_all_elements(browser, DashboardContents)
+    should_see_all_elements(browser, DashboardContentsOnSuccess)
+
+
+@pytest.mark.django_db
+@mock.patch.object(helpers, 'get_dashboard_export_opportunities')
+@mock.patch.object(helpers, 'get_dashboard_events')
+@mock.patch.object(helpers, 'create_company_profile')
+def test_dashboard_without_success_query_parameter(
+        mock_create_company_profile, mock_get_dashboard_events,
+        mock_get_dashboard_export_opportunities, mock_get_company_profile,
+        server_user_browser_dashboard, single_event, single_opportunity
+):
+    def side_effect(_):
+        mock_get_company_profile.return_value = {
+            'expertise_countries': [],
+            'expertise_industries': ['SL10001', 'SL10002'],
+        }
+
+    mock_get_dashboard_events.return_value = create_response()
+    mock_get_dashboard_events.side_effect = [[], [single_event], [single_event]]
+    mock_get_dashboard_export_opportunities.return_value = create_response()
+    mock_get_dashboard_export_opportunities.side_effect = [
+        [], [single_opportunity], [single_opportunity]
+    ]
+    mock_create_company_profile.return_value = create_response()
+    mock_create_company_profile.side_effect = side_effect
+    live_server, user, browser = server_user_browser_dashboard
+    should_not_see_errors(browser)
+
+    should_see_all_elements(browser, DashboardModalLetsGetToKnowYou)
+
+    sector_labels = [label for _, label in choices.SECTORS]
+    industries = random.sample(sector_labels, random.randint(1, 5))
+    submit_industries(browser, industries)
+
+    attach_jpg_screenshot(browser, 'Dashboard with success query parameter')
+    browser.get(f'{live_server.url}/dashboard/')
+    attach_jpg_screenshot(browser, 'Dashboard without success query parameter')
+    should_not_see(browser, DashboardModalLetsGetToKnowYou)
+    should_see_all_elements(browser, HeaderSignedIn)
+    should_see_all_elements(browser, DashboardContents)
+    should_see_all_elements(browser, DashboardContentsWithoutSuccess)